[dtensor] more refactor to imports/paths
as titled

[ghstack-poisoned]
